refactor: common/ui, modal 분리
