+++ src/common/util/index.ts
@@ -1,28 +1,3 @@
-import { confirmAlert } from "react-confirm-alert";
-import "react-confirm-alert/src/react-confirm-alert.css";
-
-type ModalProps = {
-  title: string;
-  message: string;
-  onConfirm: () => void;
-};
-
-export const handleModal = ({ title, message, onConfirm }: ModalProps) => {
-  confirmAlert({
-    title,
-    message,
-    buttons: [
-      {
-        label: "확인",
-        onClick: () => onConfirm(),
-      },
-      {
-        label: "취소",
-        onClick: () => { },
-      },
-    ],
-  });
-}
 export const printWon = (price: number): string => {
   return price.toLocaleString('en-US') + " 원";
 }
